requirements.txt
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.18.0
streamlit-extras
types-requests
types-python-dateutil
mypy
types-plotly
types-python-dotenv
mypy-extensions
pandas>=2.1.0
pandas-stubs